button and label now share code via inheritance
