+++ LightDrive/Workspace/Widgets/Desk/__init__.py
@@ -1,0 +1,2 @@
+from .button import DeskButton
+from .label import DeskLabel
